#19 #18 changed the way the API routes work, post now supported
The old way produced too much code duplication and would be too massive
and hard to maintain.

Instead, added a way to validate the command and use an argument parser
to translate the URI args into Redis-rb args.  Plus we can use a
response builder to build us a json response.

However any command can be run right now by sending it to the
endpoint, probably not a good thing.